I'm, happier with the new code then the old code, so test.py has become main.py
